[Security Solution] Removing cypress folder (#197273)
## Summary

Deleting the Cypress folder that was added in `test_serverless` as a
POC.

Currently is not used and this can create misunderstandings regarding
ownership of it.

---------

(cherry picked from commit 86e6c74f7c39a512eaa43cc025434dc6a53d55f6)

# Conflicts:
#	.buildkite/pipelines/pull_request/security_solution/defend_workflows.yml
#	.github/CODEOWNERS
