+++ x-pack/test/osquery_cypress/serverless_cli_config.ts
@@ -12,9 +12,7 @@ import { startOsqueryCypress } from './runner';
 
 export default async function ({ readConfigFile }: FtrConfigProviderContext) {
   const securitySolutionCypressConfig = await readConfigFile(
-    require.resolve(
-      '@kbn/test-suites-serverless/functional/test_suites/security/cypress/security_config.base'
-    )
+    require.resolve('./serverless_config.base.ts')
   );
 
   return {
