--- x-pack/test_serverless/functional/test_suites/security/cypress/.eslintrc.json
@@ -1,13 +1,0 @@
-{
-  "plugins": ["cypress"],
-  "extends": [
-    "plugin:cypress/recommended"
-  ],
-  "env": {
-    "cypress/globals": true
-  },
-  "rules": {
-    "cypress/no-force": "warn",
-    "import/no-extraneous-dependencies": "off"
-  }
-}
