--- x-pack/test_serverless/functional/test_suites/security/cypress/.gitignore
@@ -1,3 +1,0 @@
-videos
-screenshots
-downloads
