--- x-pack/test_serverless/functional/test_suites/security/cypress/reporter_config.json
@@ -1,10 +1,0 @@
-{
-  "reporterEnabled": "mochawesome, mocha-junit-reporter",
-  "reporterOptions": {
-    "html": false,
-    "json": true,
-    "mochaFile": "../../../../../../target/kibana-security-serverless/cypress/results/TEST-security-solution-cypress-[hash].xml",
-    "overwrite": false,
-    "reportDir": "../../../../../../target/kibana-security-serverless/cypress/results"
-  }
-}
