--- x-pack/test_serverless/functional/test_suites/security/cypress/tasks/login.ts
@@ -1,87 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { request } from '@kbn/security-solution-plugin/public/management/cypress/tasks/common';
-import { LoginState } from '@kbn/security-plugin/common/login_state';
-import type { ServerlessRoleName } from '../../../../../shared/lib';
-import { ServerlessRoleName as RoleName } from '../../../../../shared/lib/security/types';
-import { STANDARD_HTTP_HEADERS } from '../../../../../shared/lib/security/default_http_headers';
-
-/**
- * Send login via API
- * @param username
- * @param password
- *
- * @private
- */
-const sendApiLoginRequest = (
-  username: string,
-  password: string
-): Cypress.Chainable<{ username: string; password: string }> => {
-  const baseUrl = Cypress.config().baseUrl;
-
-  cy.log(`Authenticating [${username}] via ${baseUrl}`);
-
-  const headers = { ...STANDARD_HTTP_HEADERS };
-  return request<LoginState>({ headers, url: `${baseUrl}/internal/security/login_state` })
-    .then((loginState) => {
-      const basicProvider = loginState.body.selector.providers.find(
-        (provider) => provider.type === 'basic'
-      );
-      return request({
-        url: `${baseUrl}/internal/security/login`,
-        method: 'POST',
-        headers,
-        body: {
-          providerType: basicProvider?.type,
-          providerName: basicProvider?.name,
-          currentURL: '/',
-          params: { username, password },
-        },
-      });
-    })
-    .then(() => ({ username, password }));
-};
-
-interface CyLoginTask {
-  (user?: ServerlessRoleName | 'elastic'): ReturnType<typeof sendApiLoginRequest>;
-
-  /**
-   * Login using any username/password
-   * @param username
-   * @param password
-   */
-  with(username: string, password: string): ReturnType<typeof sendApiLoginRequest>;
-}
-
-/**
- * Login to Kibana using API (not login page). By default, user will be logged in using
- * the username and password defined via `KIBANA_USERNAME` and `KIBANA_PASSWORD` cypress env
- * variables.
- * @param user Defaults to `soc_manager`
- */
-export const login: CyLoginTask = (
-  user: ServerlessRoleName | 'elastic' = RoleName.SOC_MANAGER
-): ReturnType<typeof sendApiLoginRequest> => {
-  let username = Cypress.env('KIBANA_USERNAME');
-  let password = Cypress.env('KIBANA_PASSWORD');
-
-  if (user && user !== 'elastic') {
-    return cy.task('loadUserAndRole', { name: user }).then((loadedUser) => {
-      username = loadedUser.username;
-      password = loadedUser.password;
-
-      return sendApiLoginRequest(username, password);
-    });
-  } else {
-    return sendApiLoginRequest(username, password);
-  }
-};
-
-login.with = (username: string, password: string): ReturnType<typeof sendApiLoginRequest> => {
-  return sendApiLoginRequest(username, password);
-};
