--- x-pack/test_serverless/functional/test_suites/security/cypress/tasks/navigation.ts
@@ -1,10 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-export const navigatesToLandingPage = () => {
-  cy.visit('/app/security/get_started');
-};
